Add reusable CI default-ref policy check
