+++ genie/ci-workflow.ts
@@ -23,2115 +23,90 @@
  * ```
  */
 
-import {
-  githubWorkflow,
-  type ActionlintConfig,
-  type GitHubWorkflowArgs,
-} from '../packages/@overeng/genie/src/runtime/mod.ts'
-import { RUNNER_PROFILES, type RunnerProfile } from './ci.ts'
-import {
-  netlifyDeployStep as buildNetlifyDeployStep,
-  netlifyStorybookCommentStep as buildNetlifyStorybookCommentStep,
-} from './deploy-preview/netlify.ts'
-import {
-  type VercelProject,
-  vercelDeployJobs as buildVercelDeployJobs,
-  vercelDeployStep as buildVercelDeployStep,
-} from './deploy-preview/vercel.ts'
-
-export { RUNNER_PROFILES, type RunnerProfile }
-
-// =============================================================================
-// Shared Config
-// =============================================================================
-
-/** Self-hosted NixOS runner labels (x86_64-linux, e.g. dev3) */
-export const linuxX64Runner = ['sh-linux-x64', 'nix'] as const
-
-/** Self-hosted NixOS runner labels (aarch64-linux, e.g. dev4) */
-export const linuxArm64Runner = ['sh-linux-arm64', 'nix'] as const
-
-/** Self-hosted macOS runner labels (aarch64-darwin, e.g. mbp2021) */
-export const darwinArm64Runner = ['sh-darwin-arm64', 'nix'] as const
-
-/** All self-hosted runner labels — derived from the runner constants above + RUNNER_PROFILES */
-const SELF_HOSTED_RUNNER_LABELS = [
-  ...new Set([...RUNNER_PROFILES, ...linuxX64Runner, ...linuxArm64Runner, ...darwinArm64Runner]),
-] as const
-
-/** Default actionlint config with all known self-hosted runner labels */
-export const defaultActionlintConfig: ActionlintConfig = {
-  selfHostedRunnerLabels: SELF_HOSTED_RUNNER_LABELS,
-}
-
-/** Standard shell defaults for CI run steps */
-export const bashShellDefaults = {
-  run: { shell: 'bash' },
-} as const
-
-/**
- * Standard CI environment variables.
- * GITHUB_TOKEN is exported for tools that need it as a shell env var (e.g. gh CLI, nix auth).
- * Nix eval policy is enforced at step runtime by helpers like
- * `validateNixStoreStep` and `runDevenvTasksBefore`, which append
- * `restrict-eval = false` while preserving inherited NIX_CONFIG values.
- */
-export const standardCIEnv = {
-  FORCE_SETUP: '1',
-  CI: 'true',
-  GITHUB_TOKEN: '${{ github.token }}',
-} as const
-
-/**
- * Cancel superseded CI workflow runs for the same PR or branch.
- *
- * The group key intentionally does not include the job name so a new push
- * cancels the entire older workflow run rather than letting stale sibling jobs
- * continue consuming runner capacity.
- */
-export const ciWorkflowConcurrency = {
-  group: '${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}',
-  'cancel-in-progress': true,
-} as const
-
-/**
- * Standard wrapper for composed CI workflows.
- *
- * This keeps cancellation policy centralized in `effect-utils` instead of
- * making each consumer remember to wire `concurrency` by hand. Repos can still
- * override the policy by passing an explicit `concurrency` field.
- */
-export const ciWorkflow = (args: GitHubWorkflowArgs) =>
-  (({ concurrency, actionlint, ...rest }) =>
-    githubWorkflow({
-      concurrency: concurrency ?? ciWorkflowConcurrency,
-      actionlint: actionlint ?? defaultActionlintConfig,
-      ...rest,
-    }))(args)
-
-type NixConfigOptions = {
-  unrestrictedEval?: boolean
-  extraLines?: readonly string[]
-}
-
-export type NixBinaryCache = {
-  readonly uri: string
-  readonly publicKey: string
-}
-
-export const devenvBinaryCache = {
-  uri: 'https://devenv.cachix.org',
-  publicKey: 'devenv.cachix.org-1:w1cLUi8dv3hnoSPGAuibQv+f9TZLr6cv/Hm9XgU50cw=',
-} as const satisfies NixBinaryCache
-
-/** Build a binary-cache descriptor for a Cachix cache. */
-export const cachixBinaryCache = (opts: { name: string; publicKey: string }): NixBinaryCache => ({
-  uri: `https://${opts.name}.cachix.org`,
-  publicKey: opts.publicKey,
-})
-
-const dedupeBinaryCaches = (caches: readonly NixBinaryCache[]) => [
-  ...new Map(caches.map((cache) => [cache.uri, cache])).values(),
-]
-
-const cachixHostsFromBinaryCaches = (caches: readonly NixBinaryCache[]) => [
-  ...new Set(
-    caches.flatMap((cache) => {
-      const host = new URL(cache.uri).host
-      return host.endsWith('.cachix.org') ? [host] : []
-    }),
-  ),
-]
-
-/** Render `extra-conf` lines for one or more binary caches. */
-export const nixBinaryCachesExtraConf = (caches: readonly NixBinaryCache[]) => {
-  const resolvedCaches = dedupeBinaryCaches([devenvBinaryCache, ...caches])
-  return [
-    `extra-substituters = ${resolvedCaches.map((cache) => cache.uri).join(' ')}`,
-    `extra-trusted-public-keys = ${resolvedCaches.map((cache) => cache.publicKey).join(' ')}`,
-  ].join('\n')
-}
-
-const devenvBinRef = '"${DEVENV_BIN:?DEVENV_BIN not set}"'
-
-const resolveDevenvRevScript = `DEVENV_REV=$(jq -r .nodes.devenv.locked.rev devenv.lock)
-if [ -z "$DEVENV_REV" ] || [ "$DEVENV_REV" = "null" ]; then
-  echo '::error::devenv.lock missing .nodes.devenv.locked.rev'
-  exit 1
-fi`
-
-const resolveDevenvFnScript = `resolve_devenv() {
-  nix build \\
-    --accept-flake-config \\
-    --option extra-substituters ${devenvBinaryCache.uri} \\
-    --option extra-trusted-public-keys ${devenvBinaryCache.publicKey} \\
-    --no-link \\
-    --print-out-paths \\
-    "github:cachix/devenv/$DEVENV_REV#devenv"
-}`
-
-const shellSingleQuote = (value: string) => `'${value.replaceAll("'", `'"'"'`)}'`
-
-/** Build extra-conf / NIX_CONFIG content for common Nix feature flags. */
-export const nixExtraConf = (opts: NixConfigOptions = {}) =>
-  [
-    ...(opts.unrestrictedEval === true ? ['restrict-eval = false'] : []),
-    ...(opts.extraLines ?? []),
-  ].join('\n')
-
-const withAppendedNixConfig = ({
-  command,
-  opts = {},
-}: {
-  command: string
-  opts?: NixConfigOptions
-}) => {
-  const extraConf = nixExtraConf(opts)
-  if (extraConf === '') {
-    return command
-  }
-
-  const quotedExtraConf = shellSingleQuote(extraConf)
-  return `if [ -n "${'${NIX_CONFIG:-}'}" ]; then NIX_CONFIG_WITH_APPEND=$(printf '%s\\n%s' "$NIX_CONFIG" ${quotedExtraConf}); else NIX_CONFIG_WITH_APPEND=${quotedExtraConf}; fi; NIX_CONFIG="$NIX_CONFIG_WITH_APPEND" ${command}`
-}
-
-const dollar = '$'
-
-/**
- * Fall back to the standard CI pnpm paths when a workflow has not exported
- * them via `pnpmStateSetupStep` yet. This keeps `runDevenvTasksBefore` safe for
- * downstream callers while effect-utils centralizes the preferred setup step.
- */
-const withCiPnpmState = (command: string) =>
-  `PNPM_HOME="\${PNPM_HOME:-${jobLocalPnpmHome}}" PNPM_STORE_DIR="\${PNPM_STORE_DIR:-${jobLocalPnpmStore}}" ${command}`
-
-const runDevenvTasksBeforeWithOptions = (opts: NixConfigOptions, ...args: [string, ...string[]]) =>
-  withAppendedNixConfig({
-    command: withCiPnpmState(
-      `DT_PASSTHROUGH=1 ${devenvBinRef} tasks run ${args.join(' ')} --mode before`,
-    ),
-    opts,
-  })
-
-// Keep helper script bodies inline so downstream Genie imports stay bootstrap-safe.
-// The temporary import sandbox copies TypeScript modules, not arbitrary adjacent
-// shell scripts, so runtime file reads during workflow generation are unsafe.
-const nixGcRaceRetryScript = String.raw`#!/usr/bin/env bash
-
-run_nix_gc_race_retry() {
-  local task="$1"
-  local command="$2"
-  local max="${dollar}{NIX_GC_RACE_MAX_RETRIES:-10}"
-  local heartbeat="${dollar}{CI_PROGRESS_HEARTBEAT_SECONDS:-60}"
-  local attempt=1
-  local log rc path start now elapsed hb_pid flattened saw_invalid_path saw_cachix_signature had_errexit
-
-  start="$(date +%s)"
-
-  write_summary() {
-    [ -n "${dollar}{GITHUB_STEP_SUMMARY:-}" ] || return 0
-    {
-      echo "### CI Task"
-      echo "- Task: $task"
-      echo "- Status: $1"
-      echo "- Duration: $elapsed s"
-      echo "- Attempts: $attempt/$max"
-      [ -z "${dollar}{2:-}" ] || echo "- Note: $2"
-    } >> "$GITHUB_STEP_SUMMARY"
-  }
-
-  while [ "$attempt" -le "$max" ]; do
-    echo "::notice::[ci] starting $task (attempt $attempt/$max)"
-    (
-      while sleep "$heartbeat"; do
-        now=$(date +%s)
-        elapsed=$((now - start))
-        echo "::notice::[ci] $task still running after $elapsed s (attempt $attempt/$max)"
-      done
-    ) &
-    hb_pid=$!
-
-    log=$(mktemp)
-    had_errexit=false
-    case $- in
-      *e*) had_errexit=true ;;
-    esac
-    set +e
-    eval "$command" > >(tee -a "$log") 2> >(tee -a "$log" >&2)
-    rc=$?
-    if [ "$had_errexit" = true ]; then
-      set -e
-    fi
-
-    kill "$hb_pid" 2>/dev/null || true
-    wait "$hb_pid" 2>/dev/null || true
-
-    now=$(date +%s)
-    elapsed=$((now - start))
-
-    if [ "$rc" -eq 0 ]; then
-      echo "::notice::[ci] completed $task in $elapsed s"
-      if [ "$attempt" -gt 1 ]; then
-        write_summary success "Recovered from Nix GC race after retry"
-      else
-        write_summary success
-      fi
-      rm -f "$log"
-      return 0
-    fi
-
-    flattened=$(tr '\r\n' '  ' < "$log" | sed -E $'s/\x1B\[[0-9;]*m//g')
-    path=$(printf '%s' "$flattened" |
-      grep -o "error:[[:space:]]*path '/nix/store/[^']*'[[:space:]]*is not valid" |
-      head -1 |
-      grep -o "/nix/store/[^']*" |
-      tr -d '[:space:]' || true)
-    saw_invalid_path=false
-    saw_cachix_signature=false
-    [ -n "$path" ] && saw_invalid_path=true
-    printf '%s' "$flattened" | grep -Eq 'error:[[:space:]]*.*Failed to convert config\.cachix to JSON' && saw_cachix_signature=true || true
-    printf '%s' "$flattened" | grep -Eq 'error:[[:space:]]*.*while evaluating the option.*cachix\.package' && saw_cachix_signature=true || true
-    rm -f "$log"
-
-    if [ "$saw_invalid_path" != true ] && [ "$saw_cachix_signature" != true ]; then
-      echo "::warning::[ci] $task failed after $elapsed s without a detected Nix store validity race"
-      write_summary failure "No Nix GC race signature detected"
-      return "$rc"
-    fi
-
-    if [ "$saw_cachix_signature" = true ] && [ -n "$path" ]; then
-      echo "::warning::Nix store validity race detected for $task via cachix eval wrapper (attempt $attempt/$max): $path"
-    elif [ "$saw_cachix_signature" = true ]; then
-      echo "::warning::Nix store validity race detected for $task via cachix eval wrapper without extracted store path (attempt $attempt/$max)"
-    else
-      echo "::warning::Nix store validity race detected for $task (attempt $attempt/$max): $path"
-    fi
-
-    [ -z "$path" ] || nix-store --realise "$path" 2>/dev/null || true
-    rm -rf ~/.cache/nix/eval-cache-*
-    attempt=$((attempt + 1))
-  done
-
-  now=$(date +%s)
-  elapsed=$((now - start))
-  echo "::error::Nix GC race retry exhausted for $task ($max attempts)"
-  write_summary failure "Nix GC race retry exhausted"
-  return 1
-}`
-
-/**
- * Retry wrapper for the Nix store validity race where `derivationStrict` fails
- * with `path '/nix/store/...' is not valid`.
- *
- * In practice this often surfaces through higher-level eval wrappers such as
- * `Failed to convert config.cachix to JSON` / `while evaluating the option
- * cachix.package` before the final invalid-store-path line appears. We treat
- * those as the same root cause and retry after realizing the missing path and
- * clearing the eval cache.
- *
- * TODO: Remove once NixOS/nix#15469 and DeterminateSystems/nix-src#395 are released
- * @see https://github.com/NixOS/nix/pull/15469
- * @see https://github.com/DeterminateSystems/nix-src/issues/395
- */
-const withGcRaceRetry = ({ command, label }: { command: string; label: string }) => {
-  const quotedCommand = shellSingleQuote(command)
-  const quotedLabel = shellSingleQuote(label)
-  return [
-    '__nix_gc_retry_helper=$(mktemp)',
-    'cat > "$__nix_gc_retry_helper" <<\'EOF\'',
-    nixGcRaceRetryScript,
-    'EOF',
-    '. "$__nix_gc_retry_helper"',
-    'rm -f "$__nix_gc_retry_helper"',
-    `run_nix_gc_race_retry ${quotedLabel} ${quotedCommand}`,
-  ].join('\n')
-}
-
-/** Build a command that runs one or more devenv tasks with `--mode before`. */
-export const runDevenvTasksBefore = (...args: [string, ...string[]]) =>
-  withGcRaceRetry({
-    command: runDevenvTasksBeforeWithOptions({ unrestrictedEval: true }, ...args),
-    label: `devenv tasks run ${args.join(' ')} --mode before`,
-  })
-
-export type DevenvPerfProbe = {
-  readonly name: string
-  readonly command: readonly [string, ...string[]]
-  readonly traceOutput?: string
-}
-
-export type CiMeasurementObservation = {
-  readonly name: string
-  readonly unit: string
-  readonly value: number
-  readonly dimensions?: Record<string, string | number | boolean | null>
-}
-
-export const ciMeasurementMetrics = {
-  devenvProbeDuration: 'devenv.<probe>.duration',
-  nixClosureNarSize: 'nix.closure.nar_size',
-  nixClosurePathCount: 'nix.closure.path_count',
-  nixClosureBucketNarSize: 'nix.closure.bucket.nar_size',
-} as const
-
-export type NixClosureMeasurementBucket = {
-  readonly name: string
-  readonly pathRegex: string
-}
-
-export type NixClosureMeasurementStepOptions = {
-  readonly installable: string
-  readonly targetName?: string
-  readonly targetSystem?: string
-  readonly artifactDir?: string
-  readonly artifactFile?: string
-  readonly buckets?: readonly NixClosureMeasurementBucket[]
-}
-
-type DevenvPerfSetupStep = GitHubWorkflowArgs['jobs'][string]['steps'][number]
-
-export type DevenvPerfJobOptions = {
-  readonly runsOn?: readonly string[]
-  readonly artifactDir?: string
-  readonly artifactName?: string
-  readonly setupSteps?: readonly DevenvPerfSetupStep[]
-  readonly env?: Record<string, string>
-  readonly taskProbes?: readonly string[]
-  readonly probes?: readonly DevenvPerfProbe[]
-  readonly retentionDays?: number
-  readonly regressionMode?: 'off' | 'warn' | 'fail'
-}
-
-const devenvPerfProbeLine = (probe: DevenvPerfProbe) => {
-  const args = probe.command.map(shellSingleQuote).join(' ')
-  const trace = probe.traceOutput ?? ''
-  return `measure ${shellSingleQuote(probe.name)} ${shellSingleQuote(trace)} ${args}`
-}
-
-const defaultDevenvPerfTaskProbe = (task: string): DevenvPerfProbe => ({
-  name: `task_${task.replaceAll(':', '_')}`,
-  command: ['$DEVENV_BIN', 'tasks', 'run', task, '--mode', 'before', '--no-tui', '--show-output'],
-})
-
-const renderDevenvPerfScript = (
-  opts: Required<Pick<DevenvPerfJobOptions, 'taskProbes' | 'probes'>>,
-) => {
-  const probes: readonly DevenvPerfProbe[] = [
-    {
-      name: 'shell_eval_traced',
-      command: [
-        '$DEVENV_BIN',
-        '--trace-to',
-        'json:file:$trace_file',
-        'shell',
-        '--no-reload',
-        '--',
-        'true',
-      ],
-      traceOutput: '$ARTIFACT_DIR/traces/shell_eval_traced.json',
-    },
-    { name: 'shell_eval_warm', command: ['$DEVENV_BIN', 'shell', '--no-reload', '--', 'true'] },
-    { name: 'tasks_list', command: ['$DEVENV_BIN', 'tasks', 'list'] },
-    { name: 'processes_help', command: ['$DEVENV_BIN', 'processes', '--help'] },
-    ...opts.taskProbes.map(defaultDevenvPerfTaskProbe),
-    ...opts.probes,
-  ]
-
-  return String.raw`set -euo pipefail
-
-mkdir -p "$ARTIFACT_DIR/traces"
-
-{
-  printf 'timestamp_utc=%s\n' "$(date -u +%Y-%m-%dT%H:%M:%SZ)"
-  printf 'repository=%s\n' "${dollar}{GITHUB_REPOSITORY:-unknown}"
-  printf 'ref=%s\n' "${dollar}{GITHUB_REF:-unknown}"
-  printf 'sha=%s\n' "${dollar}{GITHUB_SHA:-unknown}"
-  printf 'runner_name=%s\n' "${dollar}{RUNNER_NAME:-unknown}"
-  printf 'runner_os=%s\n' "${dollar}{RUNNER_OS:-unknown}"
-  printf 'runner_arch=%s\n' "${dollar}{RUNNER_ARCH:-unknown}"
-  printf 'devenv_rev=%s\n' "${dollar}{DEVENV_REV:-unknown}"
-  printf 'otel_service_name=%s\n' "${dollar}{OTEL_SERVICE_NAME:-unknown}"
-  df -h / /nix 2>/dev/null || df -h /
-  ps -eo pid,ppid,stat,etime,pcpu,pmem,comm,args 2>/dev/null \
-    | grep -E 'devenv direnv-export|nix-daemon|nix build|nix flake|github-runner' \
-    | grep -v grep || true
-} >"$ARTIFACT_DIR/host-context.txt"
-
-printf '[' >"$ARTIFACT_DIR/timings.json"
-first=1
-
-json_append_timing() {
-  local name="$1"
-  local status="$2"
-  local duration_ms="$3"
-  local stdout="$4"
-  local stderr="$5"
-  local trace="$6"
-
-  if [ "$first" -eq 0 ]; then
-    printf ',' >>"$ARTIFACT_DIR/timings.json"
-  fi
-  first=0
-
-  jq -cn \
-    --arg name "$name" \
-    --argjson status "$status" \
-    --argjson durationMs "$duration_ms" \
-    --arg stdout "$stdout" \
-    --arg stderr "$stderr" \
-    --arg trace "$trace" \
-    '{name:$name,status:$status,durationMs:$durationMs,stdout:$stdout,stderr:$stderr,trace:(if $trace == "" then null else $trace end)}' \
-    >>"$ARTIFACT_DIR/timings.json"
-}
-
-measure() {
-  local name="$1"
-  local trace_file="$2"
-  shift 2
-  case "$trace_file" in
-    '$ARTIFACT_DIR'*) trace_file="${dollar}{ARTIFACT_DIR}${dollar}{trace_file#'$ARTIFACT_DIR'}" ;;
-  esac
-  local stdout="$ARTIFACT_DIR/$name.stdout"
-  local stderr="$ARTIFACT_DIR/$name.stderr"
-  local started ended status duration_ms
-
-  mkdir -p "$(dirname "$trace_file")"
-  started="$(date +%s%3N)"
-  set +e
-  expanded=()
-  for arg in "$@"; do
-    case "$arg" in
-      '$DEVENV_BIN') expanded+=("${dollar}{DEVENV_BIN:?DEVENV_BIN not set}") ;;
-      '$ARTIFACT_DIR'*) expanded+=("${dollar}{ARTIFACT_DIR}${dollar}{arg#'$ARTIFACT_DIR'}") ;;
-      'json:file:$trace_file') expanded+=("json:file:$trace_file") ;;
-      '$trace_file') expanded+=("file:$trace_file") ;;
-      *) expanded+=("$arg") ;;
-    esac
-  done
-  "${dollar}{expanded[@]}" >"$stdout" 2>"$stderr"
-  status=$?
-  set -e
-  ended="$(date +%s%3N)"
-  duration_ms=$((ended - started))
-
-  json_append_timing "$name" "$status" "$duration_ms" "$stdout" "$stderr" "$trace_file"
-
-  if [ "$status" -ne 0 ]; then
-    echo "::error::$name failed after ${dollar}{duration_ms}ms; stderr tail follows"
-    tail -80 "$stderr" || true
-    return "$status"
-  fi
-}
-
-${probes.map(devenvPerfProbeLine).join('\n')}
-
-printf ']\n' >>"$ARTIFACT_DIR/timings.json"
-
-jq . "$ARTIFACT_DIR/timings.json" >"$ARTIFACT_DIR/timings.pretty.json"
-jq -n \
-  --slurpfile timings "$ARTIFACT_DIR/timings.json" \
-  --arg schemaVersion "1" \
-  --arg generatedAt "$(date -u +%Y-%m-%dT%H:%M:%SZ)" \
-  --arg repository "${dollar}{GITHUB_REPOSITORY:-unknown}" \
-  --arg ref "${dollar}{GITHUB_REF:-unknown}" \
-  --arg sha "${dollar}{GITHUB_SHA:-unknown}" \
-  --arg runnerName "${dollar}{RUNNER_NAME:-unknown}" \
-  --arg runnerOs "${dollar}{RUNNER_OS:-unknown}" \
-  --arg runnerArch "${dollar}{RUNNER_ARCH:-unknown}" \
-  --arg devenvRev "${dollar}{DEVENV_REV:-unknown}" \
-  --arg otelServiceName "${dollar}{OTEL_SERVICE_NAME:-unknown}" \
-  '{
-    schemaVersion: $schemaVersion,
-    generatedAt: $generatedAt,
-    repository: $repository,
-    ref: $ref,
-    sha: $sha,
-    runner: { name: $runnerName, os: $runnerOs, arch: $runnerArch },
-    devenv: { rev: $devenvRev },
-    otel: { serviceName: $otelServiceName },
-    checks: ($timings[0] | map({ key: .name, value: . }) | from_entries)
-  }' >"$ARTIFACT_DIR/summary.json"
-
-jq -n \
-  --slurpfile timings "$ARTIFACT_DIR/timings.json" \
-  --argjson schemaVersion 1 \
-  --arg generatedAt "$(date -u +%Y-%m-%dT%H:%M:%SZ)" \
-  --arg repository "${dollar}{GITHUB_REPOSITORY:-unknown}" \
-  --arg branchKind "${dollar}{GITHUB_EVENT_NAME:-unknown}" \
-  --arg ref "${dollar}{GITHUB_REF:-unknown}" \
-  --arg headSha "${dollar}{GITHUB_SHA:-unknown}" \
-  --arg baseSha "${dollar}{GITHUB_BASE_SHA:-}" \
-  --arg runnerName "${dollar}{RUNNER_NAME:-unknown}" \
-  --arg runnerOs "${dollar}{RUNNER_OS:-unknown}" \
-  --arg runnerArch "${dollar}{RUNNER_ARCH:-unknown}" \
-  --arg runnerClass "${dollar}{RUNNER_CLASS:-unknown}" \
-  --arg githubRunId "${dollar}{GITHUB_RUN_ID:-unknown}" \
-  --arg githubRunAttempt "${dollar}{GITHUB_RUN_ATTEMPT:-unknown}" \
-  --arg githubJob "${dollar}{GITHUB_JOB:-unknown}" \
-  --arg taskId "${dollar}{CROSSTASK_TASK_ID:-}" \
-  --arg taskAttemptId "${dollar}{CROSSTASK_ATTEMPT_ID:-}" \
-  --arg traceId "${dollar}{TRACE_ID:-}" \
-  --arg devenvRev "${dollar}{DEVENV_REV:-unknown}" \
-  --arg otelServiceName "${dollar}{OTEL_SERVICE_NAME:-unknown}" \
-  --arg targetSystem "${dollar}{DEVENV_SYSTEM:-${dollar}{RUNNER_OS:-unknown}}" \
-  '{
-    schemaVersion: $schemaVersion,
-    generatedAt: $generatedAt,
-    producer: { name: "effect-utils-ci-measurement", version: 1 },
-    subject: {
-      repo: $repository,
-      branchKind: (if $branchKind == "" then "unknown" else $branchKind end),
-      ref: $ref,
-      headSha: $headSha,
-      baseSha: $baseSha
-    },
-    execution: {
-      provider: (if ($githubRunId != "" and $githubRunId != "unknown") then "github-actions" else "local" end),
-      workflow: "CI",
-      job: $githubJob,
-      runId: $githubRunId,
-      runAttempt: $githubRunAttempt,
-      taskId: $taskId,
-      attemptId: $taskAttemptId,
-      traceId: $traceId,
-      runner: { name: $runnerName, os: $runnerOs, arch: $runnerArch, class: $runnerClass }
-    },
-    target: { kind: "devenv", name: "dev-shell", system: $targetSystem },
-    observations: (
-      $timings[0]
-      | map({
-          name: ("devenv." + .name + ".duration"),
-          unit: "seconds",
-          value: (.durationMs / 1000),
-          dimensions: {
-            probe: .name,
-            status: .status,
-            devenvRev: $devenvRev,
-            otelServiceName: $otelServiceName
-          }
-        })
-    ),
-    artifacts: [
-      { name: "host-context", path: "host-context.txt", contentType: "text/plain" },
-      { name: "timings", path: "timings.json", contentType: "application/json" },
-      { name: "summary", path: "summary.json", contentType: "application/json" },
-      { name: "shell-eval-trace", path: "traces/shell_eval_traced.json", contentType: "application/json" }
-    ],
-    details: {
-      stdoutStderrByProbe: (
-        $timings[0]
-        | map({ key: .name, value: { stdout: .stdout, stderr: .stderr, trace: .trace } })
-        | from_entries
-      )
-    }
-  }' >"$ARTIFACT_DIR/measurements.json"
-
-compare_baseline() {
-  local baseline_path="${dollar}{DEVENV_PERF_BASELINE_SUMMARY:-$ARTIFACT_DIR/baseline/summary.json}"
-  local mode="${dollar}{DEVENV_PERF_REGRESSION_MODE:-warn}"
-
-  if [ "$mode" = "off" ]; then
-    jq -n --argjson schemaVersion 1 --arg status skipped --arg mode "$mode" '{schemaVersion:$schemaVersion, status:$status, mode:$mode, checks:{}}' >"$ARTIFACT_DIR/perf-comparison.json"
-    return 0
-  fi
-
-  if [ ! -f "$baseline_path" ]; then
-    jq -n \
-      --argjson schemaVersion 1 \
-      --arg status baseline_missing \
-      --arg mode "$mode" \
-      --arg baseline "$baseline_path" \
-      '{schemaVersion:$schemaVersion, status:$status, mode:$mode, baseline:$baseline, checks:{}}' \
-      >"$ARTIFACT_DIR/perf-comparison.json"
-    echo "::notice::devenv perf baseline not found at $baseline_path; recorded current measurements only"
-    return 0
-  fi
-
-  jq -n \
-    --slurpfile current "$ARTIFACT_DIR/summary.json" \
-    --slurpfile baseline "$baseline_path" \
-    --argjson schemaVersion 1 \
-    --arg mode "$mode" \
-    --arg baselinePath "$baseline_path" \
-    '
-      def budget($name):
-        if $name == "shell_eval_traced" then
-          {warnRatio:1.25, failRatio:1.5, warnMs:1500, failMs:3000}
-        elif $name == "shell_eval_warm" then
-          {warnRatio:1.5, failRatio:2.0, warnMs:500, failMs:1000}
-        elif $name == "tasks_list" or $name == "processes_help" then
-          {warnRatio:2.0, failRatio:3.0, warnMs:250, failMs:1000}
-        else
-          {warnRatio:1.5, failRatio:2.0, warnMs:1000, failMs:3000}
-        end;
-      def classify($name; $current; $baseline):
-        budget($name) as $b
-        | ($current - $baseline) as $delta
-        | (if $baseline > 0 then ($current / $baseline) else null end) as $ratio
-        | if $baseline <= 0 then "unknown"
-          elif ($delta > $b.failMs and $current > ($baseline * $b.failRatio)) then "fail"
-          elif ($delta > $b.warnMs and $current > ($baseline * $b.warnRatio)) then "warn"
-          else "pass"
-          end as $status
-        | {status:$status, currentMs:$current, baselineMs:$baseline, deltaMs:$delta, ratio:$ratio, budget:$b};
-      ($current[0].checks // {}) as $currentChecks
-      | ($baseline[0].checks // {}) as $baselineChecks
-      | (
-          $currentChecks
-          | to_entries
-          | map(
-              .key as $name
-              | .value as $current
-              | ($baselineChecks[$name] // null) as $base
-              | {
-                  key: $name,
-                  value:
-                    if $base == null then
-                      {status:"missing_baseline", currentMs:$current.durationMs}
-                    elif ($current.status != 0) then
-                      {status:"current_failed", currentMs:$current.durationMs, baselineMs:$base.durationMs}
-                    elif ($base.status != 0) then
-                      {status:"baseline_failed", currentMs:$current.durationMs, baselineMs:$base.durationMs}
-                    else
-                      classify($name; $current.durationMs; $base.durationMs)
-                    end
-                }
-            )
-          | from_entries
-        ) as $checks
-      | (
-          if any($checks[]; .status == "fail") then "fail"
-          elif any($checks[]; .status == "warn") then "warn"
-          elif any($checks[]; .status == "missing_baseline") then "partial"
-          else "pass"
-          end
-        ) as $status
-      | {schemaVersion:$schemaVersion, status:$status, mode:$mode, baseline:$baselinePath, checks:$checks}
-    ' >"$ARTIFACT_DIR/perf-comparison.json"
-
-  local status
-  status="$(jq -r '.status' "$ARTIFACT_DIR/perf-comparison.json")"
-  case "$status:$mode" in
-    fail:fail)
-      echo "::error::devenv perf regression detected"
-      jq . "$ARTIFACT_DIR/perf-comparison.json"
-      return 1
-      ;;
-    fail:*|warn:*)
-      echo "::warning::devenv perf regression threshold exceeded"
-      jq . "$ARTIFACT_DIR/perf-comparison.json"
-      ;;
-  esac
-}
-
-compare_baseline
-
-if [ -n "${dollar}{GITHUB_STEP_SUMMARY:-}" ]; then
-  {
-    echo "### Devenv perf"
-    echo ""
-    echo "| Probe | Status | Duration |"
-    echo "| --- | ---: | ---: |"
-    jq -r '.[] | "| \(.name) | \(.status) | \(.durationMs) ms |"' "$ARTIFACT_DIR/timings.json"
-    echo ""
-    echo "- Artifact directory: \`$ARTIFACT_DIR\`"
-    echo "- OTEL service: \`${dollar}{OTEL_SERVICE_NAME:-unknown}\`"
-    echo ""
-    echo "#### Regression comparison"
-    echo ""
-    if [ -f "$ARTIFACT_DIR/perf-comparison.json" ]; then
-      jq -r '["- Status: " + .status, "- Mode: " + .mode, "- Baseline: " + (.baseline // "none")] | .[]' "$ARTIFACT_DIR/perf-comparison.json"
-    fi
-  } >>"$GITHUB_STEP_SUMMARY"
-fi
-
-cat "$ARTIFACT_DIR/timings.pretty.json"
-`
-}
-
-export const devenvPerfBenchmarkStep = (
-  opts?: Pick<DevenvPerfJobOptions, 'taskProbes' | 'probes'>,
-) =>
-  ({
-    name: 'Benchmark devenv surfaces',
-    shell: 'bash',
-    run: renderDevenvPerfScript({
-      taskProbes: opts?.taskProbes ?? [],
-      probes: opts?.probes ?? [],
-    }),
-  }) as const
-
-export const devenvPerfArtifactStep = (
-  opts?: Pick<DevenvPerfJobOptions, 'artifactDir' | 'artifactName' | 'retentionDays'>,
-) =>
-  ({
-    name: 'Upload devenv perf artifacts',
-    if: 'always()',
-    uses: 'actions/upload-artifact@v4',
-    with: {
-      name:
-        opts?.artifactName ??
-        'devenv-perf-${{ github.job }}-${{ github.run_id }}-attempt-${{ github.run_attempt }}',
-      path: opts?.artifactDir ?? 'tmp/devenv-perf-ci',
-      'if-no-files-found': 'error',
-      'retention-days': opts?.retentionDays ?? 30,
-    },
-  }) as const
-
-export const nixClosureMeasurementStep = (opts: NixClosureMeasurementStepOptions) => {
-  const artifactDir = opts.artifactDir ?? 'tmp/ci-measurements'
-  const artifactFile = opts.artifactFile ?? '$ARTIFACT_DIR/measurements.json'
-  const targetName = opts.targetName ?? opts.installable
-  const buckets = JSON.stringify(opts.buckets ?? [])
-  const targetSystemAssignment =
-    opts.targetSystem === undefined
-      ? `target_system="${dollar}{DEVENV_SYSTEM:-${dollar}{RUNNER_OS:-unknown}}"`
-      : `target_system=${shellSingleQuote(opts.targetSystem)}`
-
-  return {
-    name: `Measure Nix closure: ${targetName}`,
-    shell: 'bash',
-    env: {
-      ARTIFACT_DIR: artifactDir,
-      RUNNER_CLASS: '${{ runner.os }}-${{ runner.arch }}',
-    },
-    run: String.raw`set -euo pipefail
-
-mkdir -p "$ARTIFACT_DIR"
-installable=${shellSingleQuote(opts.installable)}
-target_name=${shellSingleQuote(targetName)}
-artifact_file=${shellSingleQuote(artifactFile)}
-${targetSystemAssignment}
-
-out_path="$(nix build --no-link --print-out-paths "$installable")"
-path_info="$ARTIFACT_DIR/nix-closure-path-info.json"
-paths_file="$ARTIFACT_DIR/nix-closure-paths.json"
-
-nix path-info --recursive --json "$out_path" >"$path_info"
-jq 'to_entries | map({ path: .key, narSize: (.value.narSize // 0) })' "$path_info" >"$paths_file"
-
-jq -n \
-  --slurpfile paths "$paths_file" \
-  --argjson schemaVersion 1 \
-  --arg generatedAt "$(date -u +%Y-%m-%dT%H:%M:%SZ)" \
-  --arg repository "${dollar}{GITHUB_REPOSITORY:-unknown}" \
-  --arg branchKind "${dollar}{GITHUB_EVENT_NAME:-unknown}" \
-  --arg ref "${dollar}{GITHUB_REF:-unknown}" \
-  --arg headSha "${dollar}{GITHUB_SHA:-unknown}" \
-  --arg baseSha "${dollar}{GITHUB_BASE_SHA:-}" \
-  --arg runnerName "${dollar}{RUNNER_NAME:-unknown}" \
-  --arg runnerOs "${dollar}{RUNNER_OS:-unknown}" \
-  --arg runnerArch "${dollar}{RUNNER_ARCH:-unknown}" \
-  --arg runnerClass "${dollar}{RUNNER_CLASS:-unknown}" \
-  --arg githubRunId "${dollar}{GITHUB_RUN_ID:-unknown}" \
-  --arg githubRunAttempt "${dollar}{GITHUB_RUN_ATTEMPT:-unknown}" \
-  --arg githubJob "${dollar}{GITHUB_JOB:-unknown}" \
-  --arg taskId "${dollar}{CROSSTASK_TASK_ID:-}" \
-  --arg taskAttemptId "${dollar}{CROSSTASK_ATTEMPT_ID:-}" \
-  --arg traceId "${dollar}{TRACE_ID:-}" \
-  --arg targetName "$target_name" \
-  --arg targetSystem "$target_system" \
-  --arg outPath "$out_path" \
-  --argjson buckets ${shellSingleQuote(buckets)} \
-  '
-    ($paths[0] // []) as $closurePaths
-    | ($closurePaths | map(.narSize) | add // 0) as $totalNarSize
-    | ($closurePaths | length) as $pathCount
-    | ($buckets | map(
-        . as $bucket
-        | {
-            name: "nix.closure.bucket.nar_size",
-            unit: "bytes",
-            value: (
-              $closurePaths
-              | map(select(.path | test($bucket.pathRegex)) | .narSize)
-              | add // 0
-            ),
-            dimensions: { bucket: $bucket.name }
-          }
-      )) as $bucketObservations
-    | {
-        schemaVersion: $schemaVersion,
-        generatedAt: $generatedAt,
-        producer: { name: "effect-utils-ci-measurement", version: 1 },
-        subject: {
-          repo: $repository,
-          branchKind: (if $branchKind == "" then "unknown" else $branchKind end),
-          ref: $ref,
-          headSha: $headSha,
-          baseSha: $baseSha
-        },
-        execution: {
-          provider: (if ($githubRunId != "" and $githubRunId != "unknown") then "github-actions" else "local" end),
-          workflow: "CI",
-          job: $githubJob,
-          runId: $githubRunId,
-          runAttempt: $githubRunAttempt,
-          taskId: $taskId,
-          attemptId: $taskAttemptId,
-          traceId: $traceId,
-          runner: { name: $runnerName, os: $runnerOs, arch: $runnerArch, class: $runnerClass }
-        },
-        target: { kind: "nix-closure", name: $targetName, system: $targetSystem },
-        observations: ([
-          {
-            name: "nix.closure.nar_size",
-            unit: "bytes",
-            value: $totalNarSize,
-            dimensions: { bucket: "total" }
-          },
-          {
-            name: "nix.closure.path_count",
-            unit: "count",
-            value: $pathCount,
-            dimensions: { bucket: "total" }
-          }
-        ] + $bucketObservations),
-        artifacts: [
-          { name: "nix-closure-path-info", path: "nix-closure-path-info.json", contentType: "application/json" },
-          { name: "nix-closure-paths", path: "nix-closure-paths.json", contentType: "application/json" }
-        ],
-        details: {
-          outPath: $outPath,
-          topPaths: ($closurePaths | sort_by(.narSize) | reverse | .[:30])
-        }
-      }
-  ' >"$artifact_file"
-
-cat "$artifact_file"
-`,
-  } as const
-}
-
-export const devenvPerfJob = (opts?: DevenvPerfJobOptions) => {
-  const artifactDir = opts?.artifactDir ?? 'tmp/devenv-perf-ci'
-
-  return {
-    'runs-on': opts?.runsOn ?? linuxX64Runner,
-    defaults: bashShellDefaults,
-    env: {
-      ...standardCIEnv,
-      ARTIFACT_DIR: artifactDir,
-      OTEL_SERVICE_NAME: 'devenv-perf-ci',
-      DEVENV_PERF_REGRESSION_MODE: opts?.regressionMode ?? 'warn',
-      RUNNER_CLASS: (opts?.runsOn ?? linuxX64Runner).join(','),
-      ...opts?.env,
-    },
-    steps: [
-      ...(opts?.setupSteps ?? [
-        checkoutStep(),
-        installNixStep(),
-        preparePinnedDevenvStep,
-        validateNixStoreStep,
-      ]),
-      devenvPerfBenchmarkStep({
-        taskProbes: opts?.taskProbes,
-        probes: opts?.probes,
-      }),
-      devenvPerfArtifactStep({
-        artifactDir,
-        artifactName: opts?.artifactName,
-        retentionDays: opts?.retentionDays,
-      }),
-    ],
-  } as const
-}
-
-const evictOutPathShellLines = [
-  '      if nix path-info "$outPath" >/dev/null 2>&1; then',
-  '        echo "evicting cached: $(basename "$outPath")"',
-  '        if ! nix store delete --ignore-liveness "$outPath" >/dev/null 2>&1; then',
-  '          echo "::error::failed to evict cached pnpm-deps output: $outPath"',
-  '          exit 1',
-  '        fi',
-  '        if nix path-info "$outPath" >/dev/null 2>&1; then',
-  '          echo "::error::cached pnpm-deps output still present after eviction: $outPath"',
-  '          exit 1',
-  '        fi',
-  '      fi',
-] as const
-
-const withEachPnpmDepsDrvShellLines = ({
-  flakeRef,
-  bodyLines,
-}: {
-  flakeRef: string
-  bodyLines: readonly string[]
-}) =>
-  [
-    `targetRef=${shellSingleQuote(flakeRef)}`,
-    'entriesJson=$(mktemp)',
-    'if nix eval --json "$targetRef.passthru.depsBuildEntries" >"$entriesJson" 2>/dev/null; then',
-    "  while IFS=$'\\t' read -r attrName drv; do",
-    '    [ -n "$drv" ] || continue',
-    ...bodyLines,
-    '  done < <(jq -r \'.[] | [.attrName, (.drvPath // "")] | @tsv\' "$entriesJson")',
-    'else',
-    '  topDrv=$(nix path-info --derivation "$targetRef" 2>/dev/null || true)',
-    '  if [ -n "$topDrv" ]; then',
-    '    while IFS= read -r drv; do',
-    '      [ -n "$drv" ] || continue',
-    '      attrName=""',
-    ...bodyLines,
-    '    done < <(nix-store -qR "$topDrv" 2>/dev/null | grep "pnpm-deps-[a-z0-9-]*-v[0-9].*\\.drv$" || true)',
-    '  fi',
-    'fi',
-    'rm -f "$entriesJson"',
-  ] as const
-
-/** Evict cached pnpm-deps fixed-output outputs so CI re-derives them fresh. */
-export const evictCachedPnpmDepsStep = ({
-  flakeRef,
-  name = 'Evict cached pnpm deps',
-}: {
-  flakeRef: string
-  name?: string
-}) => ({
-  name,
-  shell: 'bash',
-  run: withEachPnpmDepsDrvShellLines({
-    flakeRef,
-    bodyLines: [
-      '    while IFS= read -r outPath; do',
-      '      [ -n "$outPath" ] || continue',
-      ...evictOutPathShellLines,
-      '    done < <(nix-store -q --outputs "$drv" 2>/dev/null || true)',
-    ],
-  }).join('\n'),
-})
-
-/**
- * Namespace runner with run ID-based affinity to prevent queue jumping.
- * Adds a run ID label so runners spawned for one workflow run
- * don't steal jobs from other runs.
- */
-export const namespaceRunner = ({
-  profile,
-  runId,
-}: {
-  profile: RunnerProfile | (string & {})
-  runId: string
-}) => [profile, `namespace-features:github.run-id=${runId}`] as const
-
-// =============================================================================
-// Step Atoms
-// =============================================================================
-
-/** Checkout repository via actions/checkout@v6 */
-export const checkoutStep = (opts?: { repository?: string; ref?: string; path?: string }) => ({
-  uses: 'actions/checkout@v6' as const,
-  ...(opts !== undefined && Object.keys(opts).length > 0 ? { with: opts } : {}),
-})
-
-/** Mint a GitHub App installation token for downstream private-repo fetches. */
-export const githubAppInstallationTokenStep = (opts: {
-  id: string
-  appId: string
-  privateKey: string
-  owner: string
-  repositories: readonly [string, ...string[]]
-  name?: string
-}) => ({
-  id: opts.id,
-  name: opts.name ?? `Mint ${opts.owner} GitHub App token`,
-  uses: 'actions/create-github-app-token@v3' as const,
-  with: {
-    'app-id': opts.appId,
-    'private-key': opts.privateKey,
-    owner: opts.owner,
-    repositories: opts.repositories.join(','),
-  },
-})
-
-/**
- * Build shell env bindings for a GitHub token.
- *
- * Use this on later run steps when self-hosted wrappers or ad hoc git/nix
- * invocations must authenticate with the minted installation token.
- */
-export const githubAccessTokenEnv = (tokenExpression: string) => ({
-  GITHUB_TOKEN: tokenExpression,
-  GH_TOKEN: tokenExpression,
-})
-
-/**
- * Attach a GitHub token env binding to an existing workflow step.
- *
- * This is the supported way to pass an installation token through later steps.
- * GitHub Actions does not allow overriding `GITHUB_*` variables via `$GITHUB_ENV`.
- */
-export const withGitHubAccessTokenEnv = <
-  TStep extends {
-    env?: Record<string, string>
-  },
->(
-  step: TStep,
-  tokenExpression: string,
-): TStep => ({
-  ...step,
-  env: {
-    ...step.env,
-    ...githubAccessTokenEnv(tokenExpression),
-  },
-})
-
-const withPrivateCachixReadAuthCommand = ({
-  command,
-  cacheHosts,
-}: {
-  command: string
-  cacheHosts: readonly string[]
-}) => {
-  if (cacheHosts.length === 0) {
-    return command
-  }
-
-  return [
-    'if [ -z "${CACHIX_AUTH_TOKEN:-}" ]; then',
-    '  echo "::error::CACHIX_AUTH_TOKEN is not set"',
-    '  exit 1',
-    'fi',
-    'cachix_netrc="$(mktemp "${RUNNER_TEMP:-/tmp}/cachix-netrc.XXXXXX")"',
-    'trap \'rm -f "$cachix_netrc"\' EXIT',
-    'chmod 600 "$cachix_netrc"',
-    `for host in ${cacheHosts.map(shellSingleQuote).join(' ')}; do`,
-    `  printf 'machine %s\\npassword %s\\n' "$host" "$CACHIX_AUTH_TOKEN" >> "$cachix_netrc"`,
-    'done',
-    'if [ -n "${NIX_CONFIG:-}" ]; then',
-    '  NIX_CONFIG_WITH_APPEND=$(printf \'%s\\n%s\' "$NIX_CONFIG" "netrc-file = $cachix_netrc")',
-    'else',
-    '  NIX_CONFIG_WITH_APPEND="netrc-file = $cachix_netrc"',
-    'fi',
-    'export NIX_CONFIG="$NIX_CONFIG_WITH_APPEND"',
-    command,
-  ].join('\n')
-}
-
-/**
- * Attach job-local Cachix read auth to a shell step.
- *
- * This keeps private cache pull auth local to the step instead of relying on
- * host-global netrc state owned by the runner image.
- */
-export const withPrivateCachixReadAuth = <
-  TStep extends {
-    run: string
-    env?: Record<string, string>
-  },
->(
-  step: TStep,
-  opts: {
-    authTokenExpression: string
-    binaryCaches: readonly NixBinaryCache[]
-  },
-): TStep => {
-  const cacheHosts = cachixHostsFromBinaryCaches(opts.binaryCaches)
-  if (cacheHosts.length === 0) {
-    return step
-  }
-
-  return {
-    ...step,
-    env: {
-      ...step.env,
-      CACHIX_AUTH_TOKEN: opts.authTokenExpression,
-    },
-    run: withPrivateCachixReadAuthCommand({
-      command: step.run,
-      cacheHosts,
-    }),
-  }
-}
-
-/**
- * Append a GitHub access token line to NIX_CONFIG for later shell steps.
- *
- * This only updates `NIX_CONFIG`. Use `withGitHubAccessTokenEnv(...)` when the
- * same token also needs to be visible to self-hosted runner wrappers or other
- * tools that read `GITHUB_TOKEN` / `GH_TOKEN` from the step environment.
- */
-export const appendGitHubAccessTokenToNixConfigStep = (opts: {
-  tokenExpression: string
-  name?: string
-}) => ({
-  name: opts.name ?? 'Export GitHub access token for Nix',
-  shell: 'bash' as const,
-  run: [
-    `token=${shellSingleQuote(opts.tokenExpression)}`,
-    'if [ -n "${NIX_CONFIG:-}" ]; then',
-    '  printf "NIX_CONFIG<<EOF\\n%s\\naccess-tokens = github.com=%s\\nEOF\\n" "$NIX_CONFIG" "$token" >> "$GITHUB_ENV"',
-    'else',
-    '  printf "NIX_CONFIG<<EOF\\naccess-tokens = github.com=%s\\nEOF\\n" "$token" >> "$GITHUB_ENV"',
-    'fi',
-  ].join('\n'),
-})
-
-/**
- * Install Nix via DeterminateSystems/determinate-nix-action@v3.
- * Includes shared binary caches and github.com access-tokens
- * by default. On self-hosted where Nix is pre-installed, this action is a no-op
- * and extra-conf is silently skipped — the runner's nix wrapper handles
- * access-tokens there by reading GITHUB_TOKEN from the environment.
- */
-export const installNixStep = (opts?: {
-  binaryCaches?: readonly NixBinaryCache[]
-  extraConf?: string
-  githubAccessTokenExpression?: string
-  summarize?: boolean
-}) => ({
-  name: 'Install Nix',
-  uses: 'DeterminateSystems/determinate-nix-action@v3' as const,
-  with: {
-    'extra-conf': [
-      /**
-       * TODO: Remove explicit experimental-features override once upstream ca-derivations issues are resolved
-       * @see https://github.com/NixOS/nix/issues/12361
-       * @see https://github.com/cachix/devenv/issues/2364
-       */
-      'experimental-features = nix-command flakes',
-      /** Trust flake-level nixConfig (e.g. additional repo-local substituters) */
-      'accept-flake-config = true',
-      nixBinaryCachesExtraConf(opts?.binaryCaches ?? []),
-      `access-tokens = github.com=${opts?.githubAccessTokenExpression ?? '${{ github.token }}'}`,
-      ...(opts?.extraConf !== undefined ? [opts.extraConf] : []),
-    ].join('\n'),
-    summarize: opts?.summarize ?? true,
-  },
-})
-
-/**
- * Provide the cachix CLI to subsequent steps from a /nix/store output.
- *
- * Must run before `cachixStep` in the same job. cachix-action's
- * `which.sync('cachix', { nothrow: true })` short-circuit then skips its
- * built-in installer, so the binary stays a /nix/store path and the runner's
- * nix profile is never mutated.
- */
-export const cachixCliBuildStep = {
-  name: 'Provide cachix CLI from nixpkgs',
-  shell: 'bash',
-  run: [
-    'set -euo pipefail',
-    'out=$(nix build --no-link --print-out-paths nixpkgs#cachix)',
-    'echo "$out/bin" >> "$GITHUB_PATH"',
-  ].join('\n'),
-} as const
-
-/** Enable a Cachix binary cache. Requires `cachixCliBuildStep` earlier in the job. */
-export const cachixStep = (opts: { name: string; authToken?: string }) => ({
-  name: 'Enable Cachix cache',
-  uses: 'cachix/cachix-action@v17' as const,
-  with: {
-    name: opts.name,
-    ...(opts.authToken !== undefined ? { authToken: opts.authToken } : {}),
-  },
-})
-
-/**
- * Prepare lock-pinned devenv metadata from devenv.lock.
- */
-export const preparePinnedDevenvStep = {
-  name: 'Use pinned devenv from lock',
-  run: `${resolveDevenvRevScript}
-echo "DEVENV_REV=$DEVENV_REV" >> "$GITHUB_ENV"
-echo "Pinned devenv rev: $DEVENV_REV"`,
-  shell: 'bash',
-} as const
-
-/**
- * Keep pnpm's hot mutable content isolated per job while still allowing cache reuse across runs.
- *
- * In the pnpm 11 + GVS configuration we use today, the effective hot state lives
- * under `PNPM_HOME`, not `PNPM_STORE_DIR`. `PNPM_HOME` must stay
- * workspace-relative because the GVS links embed absolute paths and those need
- * to stay valid for relocatable artifacts like `vercel deploy --prebuilt`.
- */
-export const jobLocalPnpmHome = '${{ github.workspace }}/.pnpm-home'
-
-/**
- * Keep pnpm's auxiliary mutable store content isolated per job.
- *
- * We still wire `PNPM_STORE_DIR` explicitly for pnpm, but the primary CI cache
- * target is `PNPM_HOME` because that is where pnpm 11 GVS keeps the reusable
- * links and metadata.
- */
-export const jobLocalPnpmStore = '${{ runner.temp }}/pnpm-store/${{ github.job }}'
-
-/**
- * Canonical pnpm CI state surface for pnpm 11 + GVS on self-hosted runners.
- *
- * `PNPM_HOME` carries the hot reusable links and metadata, while the
- * auxiliary mutable store content still lives under `PNPM_STORE_DIR`. The
- * supported cache contract restores both together under one exact key.
- */
-export const jobLocalPnpmStatePaths = [jobLocalPnpmHome, jobLocalPnpmStore].join('\n')
-
-/** Job-local CI diagnostics directory used for runner pressure snapshots and install logs. */
-export const jobLocalCiDiagnosticsDir = '${{ runner.temp }}/ci-diagnostics/${{ github.job }}'
-
-/** Workspace-local cache root for mutable Nix client cache content on CI runners. */
-export const workspaceLocalNixCacheRoot = '${{ github.workspace }}/.ci-cache'
-
-/** Default Nix cache path restored/saved by the shared CI cache helpers. */
-export const workspaceLocalNixCachePath = `${workspaceLocalNixCacheRoot}/nix`
-
-/**
- * Export the canonical CI pnpm paths once so every later shell step shares the
- * same writable store and the same workspace-relative GVS projection.
- */
-export const pnpmStateSetupStep = {
-  name: 'Isolate pnpm state',
-  shell: 'bash',
-  run: [
-    `echo "PNPM_STORE_DIR=${jobLocalPnpmStore}" >> "$GITHUB_ENV"`,
-    `echo "PNPM_HOME=${jobLocalPnpmHome}" >> "$GITHUB_ENV"`,
-  ].join('\n'),
-} as const
-
-/**
- * Export the canonical workspace-local Nix cache root so later steps share the
- * same mutable client cache surface across one CI job.
- */
-export const nixCacheSetupStep = {
-  name: 'Isolate nix cache',
-  shell: 'bash',
-  run: [
-    `mkdir -p "${workspaceLocalNixCachePath}"`,
-    `echo "XDG_CACHE_HOME=${workspaceLocalNixCacheRoot}" >> "$GITHUB_ENV"`,
-  ].join('\n'),
-} as const
-
-/**
- * Export the job-local CI diagnostics directory once so later steps can
- * collect runner pressure snapshots and install logs in one place.
- */
-export const ciDiagnosticsSetupStep = {
-  name: 'Prepare CI diagnostics',
-  shell: 'bash',
-  run: `mkdir -p "${jobLocalCiDiagnosticsDir}"
-echo "CI_DIAGNOSTICS_DIR=${jobLocalCiDiagnosticsDir}" >> "$GITHUB_ENV"`,
-} as const
-
-const runnerPressureSnapshotScript = [
-  'set -euo pipefail',
-  'mkdir -p "$CI_DIAGNOSTICS_DIR"',
-  'pressure_file="$CI_DIAGNOSTICS_DIR/runner-pressure.txt"',
-  '{',
-  '  echo "timestamp_utc=$(date -u +%Y-%m-%dT%H:%M:%SZ)"',
-  '  echo "runner_name=${RUNNER_NAME:-unknown}"',
-  '  echo "runner_os=${RUNNER_OS:-unknown}"',
-  '  echo "runner_arch=${RUNNER_ARCH:-unknown}"',
-  '  echo "github_job=${GITHUB_JOB:-unknown}"',
-  '  echo',
-  '  uptime',
-  '  echo',
-  '  if command -v free >/dev/null 2>&1; then',
-  '    free -h',
-  '  elif [ -r /proc/meminfo ]; then',
-  '    cat /proc/meminfo',
-  '  elif command -v vm_stat >/dev/null 2>&1; then',
-  '    vm_stat',
-  '    if command -v memory_pressure >/dev/null 2>&1; then',
-  '      echo',
-  '      memory_pressure || true',
-  '    fi',
-  '  else',
-  '    echo "memory stats unavailable on runner"',
-  '  fi',
-  '  echo',
-  '  if [ -r /proc/pressure/memory ]; then',
-  '    cat /proc/pressure/memory',
-  '  fi',
-  '  echo',
-  '  df -h /',
-  '  echo',
-  '  if command -v ps >/dev/null 2>&1; then',
-  '    if ps -eo pid,ppid,user,%cpu,%mem,etime,stat,comm --sort=-%cpu >/dev/null 2>&1; then',
-  '      ps -eo pid,ppid,user,%cpu,%mem,etime,stat,comm --sort=-%cpu | head -15',
-  '    else',
-  '      ps -axo pid,ppid,user,%cpu,%mem,etime,stat,comm -r | head -15',
-  '    fi',
-  '  else',
-  '    echo "ps unavailable on runner"',
-  '  fi',
-  '} | tee "$pressure_file"',
-  'if [ -n "${GITHUB_STEP_SUMMARY:-}" ]; then',
-  '  {',
-  '    echo "### Runner pressure"',
-  '    echo ""',
-  '    echo "```text"',
-  '    tail -20 "$pressure_file"',
-  '    echo "```"',
-  '  } >> "$GITHUB_STEP_SUMMARY"',
-  'fi',
-].join('\n')
-
-/**
- * Capture a quick runner pressure snapshot before the install starts.
- *
- * This does not fail the job. It gives the later failure summary and artifact
- * enough context to tell whether pnpm timed out under host pressure.
- */
-export const captureRunnerPressureStep = {
-  name: 'Capture runner pressure',
-  shell: 'bash',
-  run: runnerPressureSnapshotScript,
-} as const
-
-const pnpmInstallFailureSummaryScript = [
-  'classify_pnpm_failure() {',
-  '  local log_file="$1"',
-  '  local signature="unknown"',
-  '  local evidence=""',
-  '  if grep -Eq "ERR_PNPM_META_FETCH_FAIL|ERR_PNPM_FETCH_FAIL|Socket timeout|ECONNRESET|EAI_AGAIN" "$log_file"; then',
-  '    signature="registry/network fetch"',
-  '    evidence="$(grep -Em1 \x27ERR_PNPM_META_FETCH_FAIL|ERR_PNPM_FETCH_FAIL|Socket timeout|ECONNRESET|EAI_AGAIN\x27 "$log_file" || true)"',
-  '  elif grep -Eq "ERR_PNPM_WORKSPACE_PKG_NOT_FOUND" "$log_file"; then',
-  '    signature="workspace package mismatch"',
-  '    evidence="$(grep -Em1 \x27ERR_PNPM_WORKSPACE_PKG_NOT_FOUND\x27 "$log_file" || true)"',
-  '  fi',
-  '  if [ -n "${GITHUB_STEP_SUMMARY:-}" ]; then',
-  '    {',
-  '      echo "### pnpm install failed"',
-  '      echo ""',
-  '      echo "- Classification: $signature"',
-  '      echo "- Evidence: \\`$evidence\\`"',
-  '      echo "- Log artifact: \\`$CI_DIAGNOSTICS_DIR/pnpm-install.log\\`"',
-  '      echo ""',
-  '      echo "```text"',
-  '      tail -80 "$log_file"',
-  '      echo "```"',
-  '    } >> "$GITHUB_STEP_SUMMARY"',
-  '  fi',
-  '  echo "::warning::pnpm install failed ($signature); see $CI_DIAGNOSTICS_DIR/pnpm-install.log"',
-  '}',
-].join('\n')
-
-/**
- * Run the repo-root pnpm install while teeing the full log to the diagnostics
- * directory and summarizing failures in the job output.
- */
-export const pnpmInstallWithDiagnosticsStep = () =>
-  ({
-    name: 'Install pnpm dependencies',
-    shell: 'bash',
-    run: [
-      'set -euo pipefail',
-      'mkdir -p "$CI_DIAGNOSTICS_DIR"',
-      'log_file="$CI_DIAGNOSTICS_DIR/pnpm-install.log"',
-      'set +e',
-      '(',
-      runDevenvTasksBefore('pnpm:install'),
-      ') 2>&1 | tee "$log_file"',
-      'rc=${PIPESTATUS[0]}',
-      'set -e',
-      'if [ "$rc" -ne 0 ]; then',
-      pnpmInstallFailureSummaryScript,
-      '  classify_pnpm_failure "$log_file"',
-      'fi',
-      'exit "$rc"',
-    ].join('\n'),
-  }) as const
-
-const nixCachePrimaryKey = (keyPrefix: string, hashFilesExpression: string) =>
-  `${keyPrefix}-${'${{ runner.os }}'}-${'${{ runner.arch }}'}-${hashFilesExpression}`
-
-/**
- * Restore the shared workspace-local Nix cache before expensive eval/build work.
- *
- * The default cache authority keys off the lockfiles that affect Nix inputs and
- * repo composition. Consumers can override the key prefix or hash expression
- * when a narrower surface is more appropriate.
- */
-export const restoreNixCacheStep = (opts?: {
-  keyPrefix?: string
-  stepId?: string
-  path?: string
-  hashFilesExpression?: string
-}) => {
-  const keyPrefix = opts?.keyPrefix ?? 'nix-cache-v1'
-  const path = opts?.path ?? workspaceLocalNixCachePath
-  const hashFilesExpression =
-    opts?.hashFilesExpression ?? "${{ hashFiles('devenv.lock', 'flake.lock', 'megarepo.lock') }}"
-
-  return {
-    id: opts?.stepId ?? 'restore-nix-cache',
-    name: 'Restore nix cache',
-    uses: 'actions/cache/restore@v4' as const,
-    with: {
-      path,
-      key: nixCachePrimaryKey(keyPrefix, hashFilesExpression),
-      'restore-keys': `${keyPrefix}-${'${{ runner.os }}'}-${'${{ runner.arch }}'}-`,
-    },
-  }
-}
-
-/**
- * Save the shared workspace-local Nix cache after the main task graph runs.
- *
- * Reuses the primary key emitted by the restore step so the save path stays
- * aligned with the exact cache authority evaluated earlier in the job.
- */
-export const saveNixCacheStep = (opts?: { restoreStepId?: string; path?: string }) => {
-  const restoreStepId = opts?.restoreStepId ?? 'restore-nix-cache'
-  const path = opts?.path ?? workspaceLocalNixCachePath
-
-  return {
-    name: 'Save nix cache',
-    if: `\${{ always() && steps.${restoreStepId}.outputs.cache-primary-key != '' }}`,
-    uses: 'actions/cache/save@v4' as const,
-    with: {
-      path,
-      key: `\${{ steps.${restoreStepId}.outputs.cache-primary-key }}`,
-    },
-  }
-}
-
-const pnpmStateCachePrimaryKey = (keyPrefix: string) =>
-  `${keyPrefix}-${'${{ runner.os }}'}-${'${{ runner.arch }}'}-${"${{ hashFiles('**/pnpm-lock.yaml') }}"}`
-
-/**
- * Restore the job-local pnpm state snapshot before any install work runs.
- *
- * Live pnpm state must use exact-key semantics. Prefix fallback restore keys
- * are not part of the supported contract for mutable pnpm state because they
- * blur the authority boundary between the current lockfile graph and older
- * warmed state.
- */
-export const restorePnpmStateStep = (opts?: {
-  keyPrefix?: string
-  stepId?: string
-  path?: string
-}) => {
-  const keyPrefix = opts?.keyPrefix ?? 'pnpm-state-v1'
-  const path = opts?.path ?? jobLocalPnpmStatePaths
-
-  return {
-    id: opts?.stepId ?? 'restore-pnpm-state',
-    name: 'Restore pnpm state',
-    uses: 'actions/cache/restore@v4' as const,
-    with: {
-      path,
-      // The fetched state contents are platform-specific, so the cache must
-      // isolate both OS and CPU architecture to avoid cross-platform corruption.
-      key: pnpmStateCachePrimaryKey(keyPrefix),
-    },
-  }
-}
-
-/**
- * Save the job-local pnpm state after the main task graph runs.
- *
- * Save only after prior steps succeeded. This avoids publishing partial or
- * corrupt live state after a failed dependency preparation step.
- */
-export const savePnpmStateStep = (opts?: {
-  keyPrefix?: string
-  restoreStepId?: string
-  path?: string
-}) => {
-  const keyPrefix = opts?.keyPrefix ?? 'pnpm-state-v1'
-  const restoreStepId = opts?.restoreStepId ?? 'restore-pnpm-state'
-  const path = opts?.path ?? jobLocalPnpmStatePaths
-
-  return {
-    name: 'Save pnpm state',
-    if: `\${{ success() && steps.${restoreStepId}.outputs.cache-hit != 'true' }}`,
-    uses: 'actions/cache/save@v4' as const,
-    with: {
-      path,
-      // Reuse the same primary key expression as restore. GitHub Actions does
-      // not allow nesting `${{ ... }}` inside a fallback string of another
-      // expression, so deriving the key once in TypeScript keeps the emitted
-      // workflow expression valid.
-      key: pnpmStateCachePrimaryKey(keyPrefix),
-    },
-  }
-}
-
-/**
- * Shared self-hosted CI setup for repos that prepare a devenv workspace,
- * restore warmed mutable state, and run `pnpm:install` before the main task.
- *
- * This composes the existing step atoms into one standard contract so
- * downstream repos can delete local workflow glue instead of reassembling the
- * same Nix/pnpm/diagnostics sequence by hand.
- */
-export const standardSelfHostedPnpmCiPrepSteps = (opts?: {
-  checkout?: Parameters<typeof checkoutStep>[0]
-  installNix?: Parameters<typeof installNixStep>[0]
-  restoreNixCache?: Parameters<typeof restoreNixCacheStep>[0]
-  applyMegarepoLock?: false | Parameters<typeof applyMegarepoLockStep>[0]
-  restorePnpmState?: Parameters<typeof restorePnpmStateStep>[0]
-  includeDiagnostics?: boolean
-}) =>
-  [
-    checkoutStep(opts?.checkout),
-    installNixStep(opts?.installNix),
-    preparePinnedDevenvStep,
-    nixCacheSetupStep,
-    restoreNixCacheStep(opts?.restoreNixCache),
-    validateNixStoreStep,
-    ...(opts?.applyMegarepoLock === false ? [] : [applyMegarepoLockStep(opts?.applyMegarepoLock)]),
-    pnpmStateSetupStep,
-    ciDiagnosticsSetupStep,
-    ...(opts?.includeDiagnostics === false ? [] : [captureRunnerPressureStep]),
-    restorePnpmStateStep(opts?.restorePnpmState),
-    pnpmInstallWithDiagnosticsStep(),
-  ] as const
-
-/**
- * Shared self-hosted CI tail for repos that save warmed mutable state and keep
- * pnpm / runner diagnostics attached to the finished job.
- */
-export const standardSelfHostedPnpmCiPostSteps = (opts?: {
-  savePnpmState?: Parameters<typeof savePnpmStateStep>[0]
-  saveNixCache?: Parameters<typeof saveNixCacheStep>[0]
-  includeDiagnosticsArtifact?: boolean
-}) =>
-  [
-    savePnpmStateStep(opts?.savePnpmState),
-    saveNixCacheStep(opts?.saveNixCache),
-    ...(opts?.includeDiagnosticsArtifact === false ? [] : [ciDiagnosticsArtifactStep()]),
-  ] as const
-
-/**
- * Upload CI diagnostics captured during the pnpm install / runner-pressure
- * steps as a single artifact on failure.
- */
-export const ciDiagnosticsArtifactStep = (opts?: { if?: string; retentionDays?: number }) => ({
-  name: 'Upload CI diagnostics artifact',
-  if: opts?.if ?? "failure() && env.CI_DIAGNOSTICS_DIR != ''",
-  uses: 'actions/upload-artifact@v4' as const,
-  with: {
-    name: 'ci-diagnostics-${{ github.job }}-${{ runner.os }}-run-${{ github.run_id }}-attempt-${{ github.run_attempt }}',
-    path: '${{ env.CI_DIAGNOSTICS_DIR }}',
-    'if-no-files-found': 'ignore',
-    'retention-days': opts?.retentionDays ?? 14,
-  },
-})
-
-/**
- * Validate exported pnpm fixed-output derivations by realizing them (which
- * may substitute from Cachix), then evicting the output and rebuilding from
- * scratch.
- *
- * FOD output paths are deterministic from the declared hash. If Cachix has a
- * previously-valid output (uploaded when the hash was correct), Nix substitutes
- * it without rebuilding — even if the hash is now stale. `--rebuild` also
- * must avoid shared-daemon-store heuristics. On CI runners, `nix store delete`
- * may succeed while the out path still appears valid due to lingering roots or
- * daemon-managed store state, which makes path-visibility checks flaky.
- *
- * The fix: realize once, then use `nix build --rebuild`. Nix rebuilds the FOD
- * and compares the result to the trusted store path directly. If the declared
- * hash is stale, the rebuild/check fails with the underlying hash mismatch.
- */
-export const validateColdPnpmDepsStep = ({
-  flakeRefs,
-  name = 'Cold pnpm deps validation',
-  substituters,
-}: {
-  flakeRefs: readonly [string, ...string[]]
-  name?: string
-  substituters?: readonly string[]
-}) => ({
-  name,
-  shell: 'bash',
-  run: (() => {
-    const substituterArgs =
-      substituters === undefined || substituters.length === 0
-        ? ''
-        : ` --option substituters ${shellSingleQuote(substituters.join(' '))}`
-
-    return [
-      'set -euo pipefail',
-      `for attr in ${flakeRefs.map(shellSingleQuote).join(' ')}; do`,
-      '  echo "::group::rebuild-check $attr"',
-      '  # Step 1: Realize once (may substitute) so rebuild has a trusted output to compare against.',
-      `  nix build --no-link "$attr"${substituterArgs}`,
-      '  # Step 2: Rebuild and compare locally. This fails on stale fixed-output hashes without',
-      '  # relying on whether a shared daemon store made the prior out path disappear.',
-      `  nix build --no-link --rebuild "$attr"${substituterArgs}`,
-      '  echo "::endgroup::"',
-      'done',
-    ].join('\n')
-  })(),
-})
-
-/** Evict any cached pnpm-deps outputs below a flake target and rebuild it against cache.nixos.org only. */
-export const coldFreshNixBuildStep = ({
-  flakeRef,
-  name = 'Cold fresh Nix build',
-  extraArgs = [],
-}: {
-  flakeRef: string
-  name?: string
-  extraArgs?: readonly string[]
-}) => ({
-  name,
-  shell: 'bash',
-  run: [
-    'set -euo pipefail',
-    ...withEachPnpmDepsDrvShellLines({
-      flakeRef,
-      bodyLines: [
-        '    installable="${drv}^*"',
-        '    echo "cold-building pnpm deps: ${attrName:-$drv}"',
-        '    nix build --no-link "$installable" --option substituters "https://cache.nixos.org" || true',
-        '    while IFS= read -r outPath; do',
-        '      [ -n "$outPath" ] || continue',
-        ...evictOutPathShellLines,
-        '    done < <(nix path-info "$installable" 2>/dev/null || true)',
-        '    nix build --no-link "$installable" --option substituters "https://cache.nixos.org"',
-      ],
-    }),
-    `nix build --no-link ${shellSingleQuote(flakeRef)}${extraArgs.length === 0 ? '' : ` ${extraArgs.map(shellSingleQuote).join(' ')}`} --option substituters "https://cache.nixos.org"`,
-  ].join('\n'),
-})
-
-/**
- * Guard the pnpm dependency-prep contract against regressions that would
- * silently reintroduce package-manager self-bootstrap or implicit lockfile
- * normalization inside fixed-output builds.
- */
-export const pnpmBuilderContractStep = ({
-  builderFile = 'nix/workspace-tools/lib/mk-pnpm-deps.nix',
-  name = 'Guard pnpm builder contract',
-}: {
-  builderFile?: string
-  name?: string
-}) => ({
-  name,
-  shell: 'bash',
-  run: [
-    'set -euo pipefail',
-    `builder=${shellSingleQuote(builderFile)}`,
-    'if [ ! -f "$builder" ]; then',
-    '  echo "::error::missing pnpm deps builder: $builder"',
-    '  exit 1',
-    'fi',
-    'for required in \\',
-    "  'manage-package-manager-versions=false' \\",
-    "  'npm_config_manage_package_manager_versions=false' \\",
-    "  'frozenLockfile ? true' \\",
-    "  'pnpm install --frozen-lockfile --ignore-scripts'; do",
-    '  if ! grep -Fq -- "$required" "$builder"; then',
-    '    echo "::error::missing required pnpm builder contract fragment: $required"',
-    '    exit 1',
-    '  fi',
-    'done',
-    'for forbidden in \\',
-    "  'lockfile-only' \\",
-    "  'pnpm add pnpm@'; do",
-    '  if grep -Fq -- "$forbidden" "$builder"; then',
-    '    echo "::error::forbidden pnpm builder contract fragment present: $forbidden"',
-    '    exit 1',
-    '  fi',
-    'done',
-  ].join('\n'),
-})
-
-/** Ephemeral per-job megarepo store path scoped to the CI run/attempt/job */
-export const jobLocalMegarepoStore =
-  '${{ runner.temp }}/megarepo-store/${{ github.run_id }}/${{ github.run_attempt }}/${{ github.job }}'
-
-/**
- * Install the megarepo CLI into a job-local bin directory.
- *
- * Uses the effect-utils commit from megarepo.lock when available so setup-time
- * `mr` has the same CLI contract as the shared task module used later by
- * devenv. This avoids stale self-hosted runner profile state shadowing the
- * pinned package.
- */
-export const installMegarepoStep = {
-  name: 'Install megarepo CLI',
-  run: `EU_REV=$(jq -r '.members["effect-utils"].commit // empty' megarepo.lock 2>/dev/null || true)
-if [ -n "$EU_REV" ]; then
-  MR_REF="github:overengineeringstudio/effect-utils/$EU_REV#megarepo"
-else
-  MR_REF="github:overengineeringstudio/effect-utils#megarepo"
-fi
-
-MR_OUT=$(nix build --no-link --print-out-paths "$MR_REF")
-MR_BIN_DIR="\${RUNNER_TEMP:-/tmp}/megarepo-bin"
-mkdir -p "$MR_BIN_DIR"
-ln -sf "$MR_OUT/bin/mr" "$MR_BIN_DIR/mr"
-if [ -n "\${GITHUB_PATH:-}" ]; then
-  printf '%s\n' "$MR_BIN_DIR" >> "$GITHUB_PATH"
-else
-  export PATH="$MR_BIN_DIR:$PATH"
-fi
-"$MR_BIN_DIR/mr" --version`,
-  shell: 'bash',
-} as const
-
-/** Fetch latest refs and apply megarepo workspace. */
-export const syncMegarepoWorkspaceStep = (opts?: { skip?: string[] }) => {
-  const args = ['mr', 'fetch', '--apply']
-  const skipCsv = opts?.skip?.join(',')
-  if (skipCsv !== undefined && skipCsv !== '') args.push('--skip', shellSingleQuote(skipCsv))
-  return {
-    name: 'Sync megarepo dependencies',
-    env: { MEGAREPO_STORE: jobLocalMegarepoStore },
-    run: `mkdir -p "$MEGAREPO_STORE"
-echo "Using job-local megarepo store: $MEGAREPO_STORE"
-if [ -n "${'${GITHUB_ENV:-}'}" ]; then
-  printf 'MEGAREPO_STORE=%s\n' "$MEGAREPO_STORE" >> "$GITHUB_ENV"
-fi
-${args.join(' ')}`,
-    shell: 'bash',
-  }
-}
-
-/**
- * Sync megarepo state using the locked effect-utils commit from megarepo.lock.
- * CI must use `apply --all` so the workspace stays on the checked-in lock
- * shape instead of silently drifting to newer branch heads during job setup.
- * Resolves the CLI inline with `nix run` to avoid `nix profile install`
- * conflicts on self-hosted runners.
- */
-export const applyMegarepoLockStep = (opts?: { skip?: string[] }) => {
-  const skipCsv = opts?.skip?.join(',') ?? ''
-  const skipArgs = skipCsv === '' ? '' : `--skip ${shellSingleQuote(skipCsv)}`
-  const quotedSkipCsv = shellSingleQuote(skipCsv)
-  const exportSkipMembersScript =
-    skipCsv === ''
-      ? ''
-      : `if [ -n "${'${GITHUB_ENV:-}'}" ]; then
-  printf 'MEGAREPO_SKIP_MEMBERS=%s\n' ${quotedSkipCsv} >> "$GITHUB_ENV"
-fi`
-  return {
-    name: 'Sync megarepo dependencies',
-    env: { MEGAREPO_STORE: jobLocalMegarepoStore },
-    run: `EU_REV=$(jq -r '.members["effect-utils"].commit' megarepo.lock)
-if [ -z "$EU_REV" ] || [ "$EU_REV" = "null" ]; then
-  echo '::error::megarepo.lock missing members["effect-utils"].commit'
-  exit 1
-fi
-mkdir -p "$MEGAREPO_STORE"
-echo "Using job-local megarepo store: $MEGAREPO_STORE"
-if [ -n "${'${GITHUB_ENV:-}'}" ]; then
-  printf 'MEGAREPO_STORE=%s\n' "$MEGAREPO_STORE" >> "$GITHUB_ENV"
-fi
-${exportSkipMembersScript}
-nix run "github:overengineeringstudio/effect-utils/$EU_REV#megarepo" -- apply --all${skipArgs !== '' ? ` ${skipArgs}` : ''}`,
-    shell: 'bash',
-  }
-}
-
-/**
- * Resolve the devenv binary and do a fast store-path validity check.
- *
- * Previously ran `devenv info` (~25s) as an eager canary to detect any store
- * corruption before tasks run. Now uses `nix-store --check-validity` (~1-2s)
- * which only verifies the devenv store path itself. If the store path is
- * invalid, runs a targeted repair on just that path and re-resolves.
- *
- * Still captures diagnostics dir + runner fingerprint for #272 instrumentation.
- *
- * @see https://github.com/namespacelabs/nscloud-setup/issues/8
- * @see https://github.com/overengineeringstudio/effect-utils/issues/272
- */
-export const validateNixStoreStep = {
-  name: 'Resolve devenv',
-  run: `${resolveDevenvRevScript}
-
-${resolveDevenvFnScript}
-
-# Temporary: capture diagnostics dir for #272 root-cause analysis.
-DIAG_ROOT="${'${RUNNER_TEMP:-/tmp}'}/nix-store-diagnostics-${'${GITHUB_JOB:-job}'}-${'${RUNNER_OS:-unknown}'}-${'${GITHUB_RUN_ATTEMPT:-0}'}"
-mkdir -p "$DIAG_ROOT"
-echo "NIX_STORE_DIAGNOSTICS_DIR=$DIAG_ROOT" >> "$GITHUB_ENV"
-
-{
-  echo "timestamp_utc=$(date -u +%Y-%m-%dT%H:%M:%SZ)"
-  echo "runner_name=${'${RUNNER_NAME:-unknown}'}"
-  echo "runner_os=${'${RUNNER_OS:-unknown}'}"
-  echo "runner_arch=${'${RUNNER_ARCH:-unknown}'}"
-  echo "github_job=${'${GITHUB_JOB:-unknown}'}"
-  echo "github_run_id=${'${GITHUB_RUN_ID:-unknown}'}"
-  echo "nix_user_conf_files=${'${NIX_USER_CONF_FILES:-}'}"
-  nix --version || true
-} > "$DIAG_ROOT/environment.txt" 2>&1
-
-if ! DEVENV_OUT=$(resolve_devenv 2> >(tee "$DIAG_ROOT/resolve-devenv.log" >&2)); then
-  echo "::error::resolve_devenv failed. Last 30 lines of log:"
-  tail -30 "$DIAG_ROOT/resolve-devenv.log" || true
-  exit 1
-fi
-DEVENV_BIN="$DEVENV_OUT/bin/devenv"
-
-# Fast validity check on the devenv store path (~1-2s vs ~25s for devenv info).
-if ! nix-store --check-validity "$DEVENV_OUT" 2>/dev/null; then
-  echo "::warning::devenv store path invalid, repairing targeted path..."
-  nix-store --repair-path "$DEVENV_OUT" > "$DIAG_ROOT/nix-store-verify-repair.log" 2>&1 || true
-  rm -rf "${'${XDG_CACHE_HOME:-$HOME/.cache}'}"/nix/eval-cache-* ~/.cache/nix/eval-cache-*
-  if ! DEVENV_OUT=$(resolve_devenv 2> >(tee "$DIAG_ROOT/resolve-devenv-post-repair.log" >&2)); then
-    echo "::error::resolve_devenv failed after repair. Last 30 lines of log:"
-    tail -30 "$DIAG_ROOT/resolve-devenv-post-repair.log" || true
-    exit 1
-  fi
-  DEVENV_BIN="$DEVENV_OUT/bin/devenv"
-fi
-
-echo "DEVENV_BIN=$DEVENV_BIN" >> "$GITHUB_ENV"
-"$DEVENV_BIN" version | tee "$DIAG_ROOT/devenv-version.txt"`,
-  shell: 'bash',
-} as const
-
-/**
- * Upload diagnostics captured by `validateNixStoreStep` as a CI artifact.
- * Add this step after validation/task steps so failure-path data is retained.
- */
-export const nixDiagnosticsArtifactStep = (opts?: { if?: string; retentionDays?: number }) => ({
-  name: 'Upload Nix diagnostics artifact',
-  if: opts?.if ?? "failure() && env.NIX_STORE_DIAGNOSTICS_DIR != ''",
-  uses: 'actions/upload-artifact@v4' as const,
-  with: {
-    name: 'nix-store-diagnostics-${{ github.job }}-${{ runner.os }}-run-${{ github.run_id }}-attempt-${{ github.run_attempt }}',
-    path: '${{ env.NIX_STORE_DIAGNOSTICS_DIR }}',
-    'if-no-files-found': 'ignore',
-    'retention-days': opts?.retentionDays ?? 14,
-  },
-})
-
-/** Job-level permissions required by `deployCommentStep` to post/edit PR comments. */
-export const deployCommentPermissions = {
-  contents: 'read',
-  'pull-requests': 'write',
-} as const
-
-/** Shared mode detection script for deploy comments. Sets `label` based on event type. */
-export const deployModeScript = [
-  'if [ "${{ github.event_name }}" = "push" ] && [ "${{ github.ref }}" = "refs/heads/main" ]; then',
-  '  label="prod"',
-  'elif [ "${{ github.event_name }}" = "pull_request" ]; then',
-  '  label="PR #${{ github.event.pull_request.number }}"',
-  'else',
-  '  exit 0',
-  'fi',
-].join('\n')
-
-/**
- * Reusable step that writes a deployment summary and upserts a PR comment.
- *
- * The consuming job must include `permissions: deployCommentPermissions` (or equivalent)
- * so that `github.token` can read/write PR comments.
- *
- * The provided scripts run in order and must:
- * - `modeScript`: set `label` (or `exit 0` for unsupported events)
- * - `rowsScript`: set `rows` as markdown table rows (`| a | b |\n`)
- */
-export const deployCommentStep = (opts: {
-  summaryTitle: string
-  tableHeaders: readonly [string, string]
-  modeScript: string
-  rowsScript: string
-  noRowsMessage: string
-  commentTitle?: string
-  if?: string
-}) => ({
-  name: 'Post deploy URLs',
-  if: opts.if ?? 'always() && !cancelled()',
-  shell: 'bash' as const,
-  env: {
-    GH_TOKEN: '${{ github.token }}',
-    GH_REPO: '${{ github.repository }}',
-  },
-  run: [
-    opts.modeScript,
-    '',
-    opts.rowsScript,
-    '',
-    'if [ -z "$rows" ]; then',
-    `  echo "${opts.noRowsMessage}" >> "$GITHUB_STEP_SUMMARY"`,
-    '  exit 0',
-    'fi',
-    '',
-    '# Write job summary',
-    '{',
-    `  echo "## ${opts.summaryTitle} ($label)"`,
-    '  echo ""',
-    `  echo "| ${opts.tableHeaders[0]} | ${opts.tableHeaders[1]} |"`,
-    '  echo "| --- | --- |"',
-    '  echo -e "$rows"',
-    '} >> "$GITHUB_STEP_SUMMARY"',
-    '',
-    '# Post/update PR comment',
-    'if [ "${{ github.event_name }}" = "pull_request" ]; then',
-    '  {',
-    `    echo "## ${opts.commentTitle ?? opts.summaryTitle}"`,
-    '    echo ""',
-    `    echo "| ${opts.tableHeaders[0]} | ${opts.tableHeaders[1]} |"`,
-    '    echo "| --- | --- |"',
-    '    echo -e "$rows"',
-    '  } > /tmp/comment.md',
-    '  export NIX_CONFIG="${NIX_CONFIG:+$NIX_CONFIG$\'\\n\'}access-tokens = github.com=${GH_TOKEN}"',
-    '  nix run nixpkgs#gh -- pr comment "${{ github.event.pull_request.number }}" --body-file /tmp/comment.md --edit-last 2>/dev/null \\',
-    '    || nix run nixpkgs#gh -- pr comment "${{ github.event.pull_request.number }}" --body-file /tmp/comment.md',
-    'fi',
-  ].join('\n'),
-})
-
-/**
- * Step that dispatches `upstream-changed` repository_dispatch to a target repo.
- * Add this to upstream CI workflows so merges to main trigger downstream alignment.
- *
- * Uses `nix run nixpkgs#gh` since self-hosted runners have Nix installed.
- *
- * Requires `MEGAREPO_ALIGNMENT_TOKEN` secret (fine-grained PAT with Contents + Pull Requests write).
- */
-export const dispatchAlignmentStep = (opts: {
-  /** Target repo that receives the dispatch (e.g. 'schickling/megarepo-all') */
-  targetRepo: string
-  /** Event type sent in the dispatch (default: 'upstream-changed') */
-  eventType?: string
-}) => ({
-  name: 'Dispatch alignment to coordinator',
-  env: { GH_TOKEN: '${{ secrets.MEGAREPO_ALIGNMENT_TOKEN }}' },
-  run: [
-    `export NIX_CONFIG="${"${NIX_CONFIG:+$NIX_CONFIG$'\\n'}"}access-tokens = github.com=${'${GH_TOKEN}'}"`,
-    `printf '{"event_type":"${opts.eventType ?? 'upstream-changed'}","client_payload":{"source_repo":"%s","source_sha":"%s"}}' "${'${{ github.repository }}'}" "${'${{ github.sha }}'}" | nix run nixpkgs#gh -- api repos/${opts.targetRepo}/dispatches --input -`,
-  ].join(' && '),
-  shell: 'bash',
-})
-
-/**
- * Complete notify-alignment job definition.
- * Runs on self-hosted runner after CI passes, dispatches `upstream-changed` to the coordinator.
- */
-export const notifyAlignmentJob = (opts: {
-  targetRepo: string
-  needs: readonly string[]
-  /** Branches that trigger notification (default: main only) */
-  branches?: readonly string[]
-}) => ({
-  'runs-on': linuxX64Runner,
-  needs: [...opts.needs],
-  if: `(${(opts.branches ?? ['main']).map((b) => `github.ref == 'refs/heads/${b}'`).join(' || ')}) && github.event_name == 'push'`,
-  steps: [dispatchAlignmentStep({ targetRepo: opts.targetRepo })],
-})
-
-// =============================================================================
-// Vercel Deploy Helpers
-// =============================================================================
-
-/**
- * Deploy a single Vercel project via devenv task.
- * Prod on push-to-main/schedule/dispatch, preview on PRs.
- * Captures final/raw deploy URLs plus deploy completion time and exports them
- * to both GITHUB_ENV and GITHUB_OUTPUT.
- */
-export const vercelDeployStep = (project: { name: string; urlEnvKey?: string }) =>
-  buildVercelDeployStep(project, runDevenvTasksBefore)
-
-/**
- * Configure git author so Vercel Deployment Protection
- * associates the deploy with a team member.
- */
-export const vercelGitAuthorStep = (opts: { name: string; email: string }) => ({
-  name: 'Configure git author for Vercel',
-  shell: 'bash' as const,
-  run: [
-    `git config user.name "${opts.name}"`,
-    `git config user.email "${opts.email}"`,
-    'git commit --amend --no-edit --reset-author',
-  ].join('\n'),
-})
-
-/**
- * Generate Vercel deploy jobs and optionally a combined comment collector job.
- *
- * Returns a flat record of GitHub Actions jobs:
- * - `deploy-<name>` — one per project, runs `vercelDeployStep`, exposes structured deploy metadata
- * - `post-deploy-comment` — optional lightweight job that collects URLs from all
- *   deploy jobs and posts a stateful deploy preview comment
- *
- * The helper is deployment-mode agnostic. The unified `vercel.nix` task module
- * decides whether a project runs build mode or static mode based on `cwd` vs
- * `staticDir`; CI only needs to invoke `vercel:deploy:<name>`.
- */
-export const vercelDeployJobs = (opts: {
-  projects: readonly VercelProject[]
-  /** CI job names that deploy jobs depend on */
-  needs?: readonly string[]
-  runner: readonly string[]
-  baseSteps: readonly Record<string, unknown>[]
-  env: Record<string, string>
-  /** Extra steps to add after deploy */
-  extraSteps?: readonly Record<string, unknown>[]
-  /** Deploy condition override. Default: always after CI passes, or directly on schedule. */
-  deployCondition?: string
-  /** Whether to add a combined deploy comment job. Default: true. */
-  includeComment?: boolean
-  commentTitle?: string
-  noRowsMessage?: string
-  deployStepDecorator?: (
-    step: Record<string, unknown>,
-    project: VercelProject,
-  ) => Record<string, unknown>
-}): Record<string, Record<string, unknown>> => {
-  return buildVercelDeployJobs({
-    ...opts,
-    runDevenvTasksBefore,
-    deployModeScript,
-    deployCommentPermissions,
-    bashShellDefaults,
-    commentRunner: linuxX64Runner,
-  })
-}
-
-// =============================================================================
-// Netlify Deploy Helpers
-// =============================================================================
-
-/**
- * Deploy step for Netlify storybooks via devenv tasks.
- * Runs `netlify:deploy` with prod/PR mode based on the event trigger.
- * Gracefully skips if NETLIFY_AUTH_TOKEN is not available.
- */
-export const netlifyDeployStep = () => buildNetlifyDeployStep(runDevenvTasksBefore)
-
-/**
- * Combined deploy comment step for Netlify storybook previews.
- *
- * When `packages` is provided, constructs preview URLs directly from the known package list.
- * This works regardless of whether the deploy task emits metadata markers, making it suitable
- * for repos that pin an older effect-utils version for Nix while using the latest for genie.
- *
- * When omitted, uses the metadata-based approach that reads deploy output from the deploy step.
- *
- * The `packages` shape matches the Nix `taskModules.netlify` / `taskModules.storybook` config:
- * `{ path: "flakes/oi", name: "flakes-oi" }` where `name` is the Netlify deploy alias.
- */
-export const netlifyStorybookCommentStep = (
-  site: string,
-  opts?: { packages?: ReadonlyArray<{ path: string; name: string }> },
-) => {
-  if (!opts?.packages) {
-    return buildNetlifyStorybookCommentStep(site, deployModeScript)
-  }
-
-  return deployCommentStep({
-    summaryTitle: 'Storybook Previews',
-    tableHeaders: ['Package', 'URL'],
-    noRowsMessage: 'No storybooks were deployed.',
-    modeScript: [
-      `site="${site}"`,
-      deployModeScript,
-      '# Set Netlify branch-deploy suffix based on mode',
-      'if [ "$label" = "prod" ]; then suffix=""; else suffix="-pr-${{ github.event.pull_request.number }}"; fi',
-    ].join('\n'),
-    rowsScript: [
-      'rows=""',
-      ...opts.packages.map((pkg) =>
-        [
-          `if [ -d "${pkg.path}/storybook-static" ]; then`,
-          `  rows="\${rows}| ${pkg.name} | https://${pkg.name}\${suffix}--\${site}.netlify.app |\\n"`,
-          'fi',
-        ].join('\n'),
-      ),
-    ].join('\n'),
-  })
-}
+export {
+  RUNNER_PROFILES,
+  bashShellDefaults,
+  cachixBinaryCache,
+  ciWorkflow,
+  ciWorkflowConcurrency,
+  darwinArm64Runner,
+  defaultActionlintConfig,
+  devenvBinaryCache,
+  jobLocalCiDiagnosticsDir,
+  jobLocalPnpmHome,
+  jobLocalPnpmStatePaths,
+  jobLocalPnpmStore,
+  linuxArm64Runner,
+  linuxX64Runner,
+  nixBinaryCachesExtraConf,
+  nixExtraConf,
+  runDevenvTasksBefore,
+  standardCIEnv,
+  workspaceLocalNixCachePath,
+  workspaceLocalNixCacheRoot,
+  type NixBinaryCache,
+  type RunnerProfile,
+} from './ci-workflow/shared.ts'
+export {
+  ciMeasurementMetrics,
+  devenvPerfArtifactStep,
+  devenvPerfBenchmarkStep,
+  devenvPerfJob,
+  nixClosureMeasurementStep,
+  type CiMeasurementObservation,
+  type DevenvPerfJobOptions,
+  type DevenvPerfProbe,
+  type NixClosureMeasurementBucket,
+  type NixClosureMeasurementStepOptions,
+} from './ci-workflow/measurements.ts'
+export {
+  appendGitHubAccessTokenToNixConfigStep,
+  cachixCliBuildStep,
+  cachixStep,
+  checkoutStep,
+  ciDiagnosticsArtifactStep,
+  ciDiagnosticsSetupStep,
+  captureRunnerPressureStep,
+  coldFreshNixBuildStep,
+  evictCachedPnpmDepsStep,
+  githubAccessTokenEnv,
+  githubAppInstallationTokenStep,
+  installNixStep,
+  namespaceRunner,
+  nixCacheSetupStep,
+  nixDiagnosticsArtifactStep,
+  pnpmBuilderContractStep,
+  pnpmInstallWithDiagnosticsStep,
+  pnpmStateSetupStep,
+  preparePinnedDevenvStep,
+  restoreNixCacheStep,
+  restorePnpmStateStep,
+  saveNixCacheStep,
+  savePnpmStateStep,
+  standardSelfHostedPnpmCiPostSteps,
+  standardSelfHostedPnpmCiPrepSteps,
+  validateColdPnpmDepsStep,
+  validateNixStoreStep,
+  withGitHubAccessTokenEnv,
+  withPrivateCachixReadAuth,
+} from './ci-workflow/setup.ts'
+export {
+  applyMegarepoLockStep,
+  defaultRefPolicyCheckStep,
+  installMegarepoStep,
+  jobLocalMegarepoStore,
+  syncMegarepoWorkspaceStep,
+  type DefaultRefPolicyCheckStepOptions,
+} from './ci-workflow/megarepo.ts'
+export {
+  deployCommentPermissions,
+  deployCommentStep,
+  deployModeScript,
+  dispatchAlignmentStep,
+  netlifyDeployStep,
+  netlifyStorybookCommentStep,
+  notifyAlignmentJob,
+  vercelDeployJobs,
+  vercelDeployStep,
+  vercelGitAuthorStep,
+} from './ci-workflow/deploy.ts'
